only show for beats
