+++ src/plugins/home/server/services/tutorials/lib/tutorials_registry_types.ts
@@ -8,6 +8,8 @@
 
 import type { KibanaRequest } from 'src/core/server';
 import type { TutorialSchema } from './tutorial_schema';
+export { TutorialsCategory } from '../../../../common/constants';
+
 export type {
   TutorialSchema,
   ArtifactsSchema,
@@ -19,13 +21,6 @@ export type {
   Instruction,
 } from './tutorial_schema';
 
-/** @public */
-export enum TutorialsCategory {
-  LOGGING = 'logging',
-  SECURITY_SOLUTION = 'security',
-  METRICS = 'metrics',
-  OTHER = 'other',
-}
 export type Platform = 'WINDOWS' | 'OSX' | 'DEB' | 'RPM';
 
 export interface TutorialContext {
